Move remaining constants from constants.ts to log.ts & doc/createNode.ts
Also simplifies log level parsing.
