+++ src/log.ts
@@ -1,11 +1,11 @@
-import { LogLevel, LogLevelId } from './constants.js'
+export type LogLevelId = 'silent' | 'error' | 'warn' | 'debug'
 
 export function debug(logLevel: LogLevelId, ...messages: any[]) {
-  if (LogLevel.indexOf(logLevel) >= LogLevel.DEBUG) console.log(...messages)
+  if (logLevel === 'debug') console.log(...messages)
 }
 
 export function warn(logLevel: LogLevelId, warning: string | Error) {
-  if (LogLevel.indexOf(logLevel) >= LogLevel.WARN) {
+  if (logLevel === 'debug' || logLevel === 'warn') {
     if (typeof process !== 'undefined' && process.emitWarning)
       process.emitWarning(warning)
     else console.warn(warning)
